added a copy of Akka's Logging.scala, with an Await call removed
This makes it not fail in discrete event because of Awaiting something that by definition must happen after the current Event has finished.
It also means that we do not detect failures in logging initialisation.
TODO: depend on a fork of Akka for the Logging change, rather than copying its source.